+++ src/main/java/com/dangeboer/raindream/service/AuthenticationService.java
@@ -6,6 +6,6 @@ import com.dangeboer.raindream.model.dto.LoginResponse;
 import com.dangeboer.raindream.model.entity.User;
 
 public interface AuthenticationService extends IService<User> {
-    Integer register(String username, String password, String phone);
+    Long register(String username, String password, String phone);
     LoginResponse login(String username, String password);
 }
